✅ [Test] : Mock Endoer, Decoder,  생성
