--- OpenMarket/OpenMarketTests/Models/StubItems.swift
@@ -1,11 +1,0 @@
-//
-//  StubItems.swift
-//  OpenMarketTests
-//
-//  Created by Fezravien on 2021/09/08.
-//
-
-import Foundation
-
-struct StubMarketItems: Decodable { }
-struct StubMarketItem: Decodable { }
